requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cycler==0.11.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.1.7
django-allauth==0.54.0
django-bootstrap==0.2.4
django-ckeditor==6.5.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.0
numpy==1.24.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
pycparser==2.21
pygame==2.1.3
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3
railway==0.0.4
requests==2.30.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.3
style==1.1.0
typing_extensions==4.6.3
update==0.0.1
uritemplate==4.1.1
urllib3==2.0.2
whitenoise==6.4.0
drf-spectacular==0.26.2